Move templated sequence to separate react component
