+++ client/src/components/Pages/Structure/Input/StructCompInputProps.tsx
@@ -1,0 +1,6 @@
+export interface StructuralComponentInputProps {
+  index: number,
+  id: string,
+  handleSave: CallableFunction,
+  handleCancel: CallableFunction
+}
